+++ packages/sdk/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Refactor BodyEditor and enhance selection handling. (`d28dd6b7`)
 - Refactor shortcuts handling and update UI components. (`a85cf446`)
 - Enhance response handling and UI features. (`b076b194`)